feature: Add init and run command for miniDocker
Now, this commit provides two commands for users:

1. run: run the container
2. init: initialize the container process, never called by users, it
   is a trick.

When user calls `run` function, it firsts use `/self/proc/self` to start
`miniDocker` and passes the arguments `init` for calling init commands and
also parse the arguments of `run` command to `init` command. Well, I don't
think this is a nice design.

And the other part is baisc operation.