feature: add support for cgroup
The logic for setting up the crgoup is not difficult. Just create the
corresponding directroy in the virtual file system.

For the last commit, we just pass the strings to the child as the
argument. In this commit, we use pipe to pass the arguments. It is not
difficult too.

However, writing unit tests is important. I write more complicated tests
than the book. Enjoy it.