Finish daemon container running and list the daemon containers
For now, we just exit the parent process to make the child to be orphan.
However, this is an ugly way now. Because we would have a lot of problem.
In the current logic, we wait for child to terminate in order to delete
the cgroups and workspaces. However, now if parent process exits, how
should we do that? I use two TODO to solve later.

How to list the daemon containers, because it detaches. So we need the
persistent storage. Simple idea.